+++ src/components/BottomNavigation/index.ts
@@ -1,3 +1,0 @@
-import { BottomNavigation } from "./BottomNavigation";
-
-export { BottomNavigation };