+++ src/index.ts
@@ -1,0 +1,3 @@
+export { BottomNavigation } from './components/BottomNavigation/BottomNavigation';
+export type { BottomNavigationProps, BottomNavigationItemType } from './components/BottomNavigation/BottomNavigation';
+